% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withLoader.R
\name{withLoader}
\alias{withLoader}
\title{Add a loader to a Shiny UI element}
\usage{
withLoader(ui_element, loader = "biocspin", proxy.height = "400px")
}
\arguments{
\item{ui_element}{A Shiny UI element (e.g., \code{uiOutput}, \code{plotOutput}, etc.)
for which to wrap with the loader.}

\item{loader}{\code{character(1)} A string specifying the loader type. Default is
"biocspin".}

\item{proxy.height}{\code{character(1)} A string specifying the height of the
proxy element. Default is "400px". If the \code{ui_element} already has a
height specified, this will be set to "100\%".}
}
\description{
Add a loader to a Shiny UI element
}
\examples{
\dontrun{
withLoader(plotOutput("myplot"))
}
}
\references{
\code{shinycustomloader} https://github.com/emitanaka/shinycustomloader

\verb{dna spin} https://codepen.io/knut/pen/ZopMWW
}
